Modify: D/L details VM; Add: temporary Visualize VM (in proress)

+++ src/app/src/main/java/com/hcmus/group14/moneytor/data/model/DebtLend.java
@@ -33,7 +33,7 @@ public class DebtLend implements Serializable {
     @ColumnInfo(name = "category") @Nullable
     private String category;
     @ColumnInfo(name = "value")
-    private int value;
+    private long value;
     @ColumnInfo(name = "target")
     private int target;
     @ColumnInfo(name = "debt")
@@ -76,11 +76,11 @@ public class DebtLend implements Serializable {
     {
         this.category = category;
     }
-    public int getValue()
+    public long getValue()
     {
         return value;
     }
-    public void setValue(int value)
+    public void setValue(long value)
     {
         this.value = value;
     }
